+++ modules/core/src/main/java/org/openlmis/core/domain/FacilityType.java
@@ -31,7 +31,7 @@ public class FacilityType extends BaseModel{
     this.code = code;
   }
 
-  public FacilityType(Integer id) {
+  public FacilityType(Long id) {
     this.id = id;
   }
 }
