+++ modules/core/src/main/java/org/openlmis/core/domain/SupervisoryNode.java
@@ -33,7 +33,7 @@ public class SupervisoryNode extends BaseModel implements Importable {
   @ImportField(name = "Facility Code", mandatory = true, nested = "code")
   private Facility facility;
 
-  public SupervisoryNode(Integer id) {
+  public SupervisoryNode(Long id) {
     this.id = id;
   }
