+++ modules/core/src/main/java/org/openlmis/core/repository/mapper/FacilityMapper.java
@@ -40,26 +40,26 @@ public interface FacilityMapper {
   @Select("SELECT * FROM users U, facilities F " +
     "WHERE U.facilityId = F.id AND U.id = #{userId} AND f.active = true")
   @Results(value = {@Result(property = "id", column = "facilityId")})
-  Facility getHomeFacility(Integer userId);
+  Facility getHomeFacility(Long userId);
 
   @Select("SELECT * FROM facility_types ORDER BY displayOrder")
   List<FacilityType> getAllTypes();
 
   @Select("SELECT * FROM facility_types where id = #{id}")
-  public FacilityType getFacilityTypeById(Integer id);
+  public FacilityType getFacilityTypeById(Long id);
 
   @Select("SELECT * FROM facility_operators ORDER BY displayOrder")
   List<FacilityOperator> getAllOperators();
 
   @Select("SELECT * FROM facility_operators WHERE id = #{id}")
-  public FacilityOperator getFacilityOperatorById(Integer id);
+  public FacilityOperator getFacilityOperatorById(Long id);
 
   @Select("SELECT code FROM facility_operators where id = #{id}")
   @SuppressWarnings("unused")
-  public String getFacilityOperatorCodeFor(Integer id);
+  public String getFacilityOperatorCodeFor(Long id);
 
   @Select("SELECT id FROM facility_operators where LOWER(code) = LOWER(#{code})")
-  Integer getOperatedByIdForCode(String code);
+  Long getOperatedByIdForCode(String code);
 
 
   @Select("SELECT * FROM facilities WHERE id = #{id}")
@@ -69,7 +69,7 @@ public interface FacilityMapper {
     @Result(property = "facilityType", column = "typeId", javaType = Integer.class, one = @One(select = "getFacilityTypeById")),
     @Result(property = "operatedBy", column = "operatedById", javaType = Integer.class, one = @One(select = "getFacilityOperatorById"))
   })
-  Facility getById(Integer id);
+  Facility getById(Long id);
 
   @Update("UPDATE facilities SET code = #{code}, name = #{name}, description = #{description}, gln = #{gln}," +
     "mainPhone = #{mainPhone}, fax = #{fax}, address1 = #{address1}," +
@@ -92,7 +92,7 @@ public interface FacilityMapper {
   void updateDataReportableAndActiveFor(Facility facility);
 
   @Select("SELECT id FROM facilities WHERE LOWER(code) = LOWER(#{code})")
-  Integer getIdForCode(String code);
+  Long getIdForCode(String code);
 
   @Select("SELECT DISTINCT f.* FROM facilities f " +
     "INNER JOIN programs_supported ps ON f.id=ps.facilityId " +
@@ -106,7 +106,7 @@ public interface FacilityMapper {
     @Result(property = "facilityType", column = "typeId", javaType = Integer.class, one = @One(select = "getFacilityTypeById")),
     @Result(property = "operatedBy", column = "operatedById", javaType = Integer.class, one = @One(select = "getFacilityOperatorById"))
   })
-  List<Facility> getFacilitiesBy(@Param(value = "programId") Integer programId, @Param(value = "requisitionGroupIds") String requisitionGroupIds);
+  List<Facility> getFacilitiesBy(@Param(value = "programId") Long programId, @Param(value = "requisitionGroupIds") String requisitionGroupIds);
 
   @Select("SELECT id, code, name FROM facilities WHERE " +
     "LOWER(code) LIKE '%' || LOWER(#{searchParam}) || '%' " +
@@ -122,7 +122,7 @@ public interface FacilityMapper {
     @Result(property = "facilityType", column = "typeId", javaType = Integer.class, one = @One(select = "getFacilityTypeById")),
     @Result(property = "operatedBy", column = "operatedById", javaType = Integer.class, one = @One(select = "getFacilityOperatorById"))
   })
-  Facility getHomeFacilityWithRights(@Param("userId") Integer userId, @Param("commaSeparatedRights") String commaSeparatedRights);
+  Facility getHomeFacilityWithRights(@Param("userId") Long userId, @Param("commaSeparatedRights") String commaSeparatedRights);
 
   @Select({"SELECT DISTINCT f.* FROM facilities f",
     "INNER JOIN requisition_group_members rgm ON f.id= rgm.facilityId",
